refactor(CA-877): extract pure math helpers into calculator_math.dart
Move _computePitch, _computeFence, and their 8 supporting constants
out of CalculatorBloc into a sibling part file, trimming ~55 lines
from the bloc and keeping the math grouped for clarity.
